==> 03-model.ipynb <==
## Import libraries

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.decomposition import TruncatedSVD
from sklearn.preprocessing import StandardScaler
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics.pairwise import pairwise_distances, cosine_similarity

df = pd.read_csv('./Data/no_multiples.csv')

df.head()

df.shape

df.dtypes

df.columns

### I will save the uri column, since i need them to search and add the songs to a playlist on spotify

df_id = df.drop(columns=['tempo', 'danceability', 'energy', 'key', 'loudness', 'mode',
       'speechiness', 'acousticness', 'instrumentalness', 'liveness',
       'valence', 'time_signature', 'duration_ms'])


df_id.set_index('index',inplace=True)

df_id.head()

df_id.to_csv('./Data/uri.csv', index= True)

### on my main data frame, I will set column index, which I had created earlier, as my index, and drop the URI column, so I only have numeric features

***now that I only have numeric features, I can standard scaler my data; this assures that each column will have a μ = 0 and σ = 1, in simple English it means they are now all on the same scale!*** 

df.set_index('index', inplace=True)
df.drop(columns=['uri'],inplace=True)

df.head()

ss = StandardScaler()
data = ss.fit_transform(df)

#### now I will create a dataframe with the transformmed features

df= pd.DataFrame(data, columns =df.columns, index = df.index)
df.shape

df.head(1)

### I will save this dataframe as well, since this is where I will be looking for to create my playlist based on tempo

df.to_csv('./Data/identifier.csv', index= True)

#### now I'm going to use cosine similarity from sklearn; this is a measure of similarity between two non-zero vectors, by measuring the cosign of the angle between them. As we know, the cosine of 0° and 180° are 1 and -1, respectively, and the cosine of any other angle falls between the two, with the cosine of 90° and 270° are both 0. therefore, we are comparing the position of each item along the circle.
-  Identical vectors will be orientated on top of one another, with a cosine similarity of 1 (angle of 0°).
-   A 90° angle between two vectors will be indicated with a cosine similarity of 0. 
- And a cosine similarity of -1 means two vectors are on the opposite sides of the spectrum, and will therefore have a 180° angle between them. 
#### for this recommendation system, we are assuming that each song is it's own vector, with unique features. We will create a recommendation system by comparing each song (vector) to every other song in a big matrix. 

cs = cosine_similarity(df, df)

matrix = pd.DataFrame(cs, columns=df.index, index=df.index)

matrix.head(1)

***as you can see the first song in my dataframe has a cosine similarity of 1 with itself, which proves that it is identical to itself!***

### now I need a function to take any song of choice and return 20 closest songs; I will sort them by tempo because it just makes sense; the first ten songs will be in ascending order of tempo,  and the next ten in descending. *a playlist is like a roller coaster, if it goes up, it must come down, or I'm not riding!*

def play_list(choice):
    list_of_recoms =[]
    uri= []
    
    
    for song in df.loc[df.index.str.contains(choice)].sort_values(by='tempo').index:
        list_of_recoms.append(matrix[song].sort_values(ascending = False)[0:10])
        list_of_recoms.append(matrix[song].sort_values(ascending = False)[19:9:-1])        
        recom= pd.DataFrame(list_of_recoms).T
        recom= recom.drop(columns=choice)
        return recom

play_list('harvest moon, neil young')
import dataframe_image as dfi
dfi.export(play_list('harvest moon, neil young'),"./Assets/recomendation.png")

#### lastly I will save my recomender matrix as a dataframe so I can create playlists on top of it 

matrix.to_csv('./Data/recommender.csv', index= True)
