==> 04- create_playlists.ipynb <==
## Import libraries

import os
import json
import spotipy
import requests
import numpy as np
import pandas as pd
import spotipy.util as util
from spotipy.oauth2 import SpotifyOAuth, SpotifyClientCredentials
from sklearn.preprocessing import StandardScaler
from sklearn.metrics.pairwise import pairwise_distances, cosine_similarity

### I will read in the data frame I have built my model on, and the one I use for grouping by tempo, and while I'm at it, I will read in the data frame of the names and the URI we had created in the modeling process so we can Identify the songs *doing so, I will not have to tinker with the API, and it's a HUGE detour*

### I will set all of the indexes as col=index

matrix = pd.read_csv('./Data/recommender.csv')
df= pd.read_csv('./Data/identifier.csv')
df_id= pd.read_csv('./Data/uri.csv')

df_id.set_index('index',inplace=True)
df_id.head(1)

df.set_index('index',inplace=True)
df.head(1)

matrix.set_index('index',inplace=True)

matrix.head()

### I will once again request a token from Spotify, this is just like the first time we requested a token, but this time we need to change the scope to ***playlist-modify-public*** 
***make sure you set your token to a global variable so you can use it in the next step***
**if you want your playlist to be private chance public to private here, and True to False, in the request  when clreating your playlist**

def get_token(scope=None):
    
    redirect_uri = os.environ['SPOTIPY_REDIRECT_URI']
    username = os.environ['SPOTIPY_USERNAME']
    client_id = os.environ['SPOTIPY_CLIENT_ID']
    client_secret = os.environ['SPOTIPY_CLIENT_SECRET']
    token = util.prompt_for_user_token(username, scope, client_id, client_secret, redirect_uri)
    return token

token = get_token(scope='playlist-modify-public')
sp = spotipy.Spotify(auth=token)

get_token()

### since I can't spell (dyslexic af), I will look for the song's here based on a couple of words in the title or artist's name that I can spell! 

df_id[df_id.index.str.contains('boogie')]

### now, I will run the function where I can get recommendations based on the song I choose. My output will return the URI of recommended songs.
***(set of URI's of the songs that are recommended based on the song of choice)***
### make sure you set your output to a global variable so we can use it to dump traks in the playlist!

def play_list(choice= None):
    list_of_recoms =[]
    uri= []
    for song in df.loc[df.index.str.contains(choice)].sort_values(by='tempo', ascending = True).index:
        list_of_recoms.append(matrix[song].sort_values(ascending = False)[0:10])
        list_of_recoms.append(matrix[song].sort_values(ascending = False)[19:9:-1])        
        recom= pd.DataFrame(list_of_recoms).T
        recom= recom.drop(columns=choice)
        uri_df = pd.merge(recom,df_id, left_index=True, right_index=True)
        uris= uri_df.reset_index(drop=True)
        uris= uri_df.set_index('uri')
        uri.append(uris.index)
        return uri[0].values
    
    
uri_request= play_list(choice='boogie with stu - 1993 remaster, led zeppelin')

### now, using the token we had received from Spotify, we will make a playlist; we will receive a playlist Id which we need to set to a global variable, so we can access the playlist and dump music in it!

def create_playlist(song= None):
    endpoint_url = f"https://api.spotify.com/v1/users/{os.environ['SPOTIPY_USERNAME']}/playlists"

    request_body = json.dumps({"name": f" {song}",
                                "description": f"play list based on {song}",
                                "public": True })
    response = requests.post(url = endpoint_url, data = request_body, headers={"Content-Type":"application/json", 
                        "Authorization":f"Bearer {token}"})

    url = response.json()['external_urls']['spotify']
    playlist_id = response.json()['id']
    print(f'PLAYLIST RESPONSE STATUS: {response.status_code}')
    print(f'PLAYLIST ID: {playlist_id}')
    print(f'URL: {url}')
    return playlist_id
playlist_id_returned= create_playlist('boogie with stu')

#### here is the playlist based on *boogie with stu - 1993 remaster, led zeppelin* follow that link if you're curious to see how the playlists look like

### lastly using the Spotipy library, we need to call on our username, the playlist_id and the list of recomended uri's that our recomender has returned to dump music in our playlist! 

sp.user_playlist_add_tracks(user=os.environ['SPOTIPY_USERNAME'], playlist_id=playlist_id_returned, tracks=uri_request)

# et voila!
